added IT test

+++ src/main/java/com/oleksii/stockservice/repository/StockRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockRepository extends PagingAndSortingRepository<Stock, Integer> {
+public interface StockRepository extends PagingAndSortingRepository<Stock, Long> {
 }
